Salinity model as used by Qiu and Wan (2013)

Salinity model as used by Qiu and Wan (2013)

SalMod(Q.total, SITE = "Val-I75", a.val = NULL, b.val = NULL, Q.thres = NULL)

Arguments

Q.total
total discharge from freshwater inputs. For Caloosahatchee Estuary this includes tidal basin discharge and

SITE
Current mointoring location with estalished salinity relationships. Sites include S79, BR31, Val-I75 and FtMyers. See Qiu and Qan (2013).

a.val
alpha value model coefficient

b.val
beta value model coefficient

Q.thres
discharge threshold

Value

modeled salinity based on model coefficients and total freshwater discharge

References

Qiu C, Wan Y (2013) Time series modeling and prediction of salinity in the Caloosahatchee River Estuary. Water Resources Research 49:5804–5816. doi: 10.1002/wrcr.20415

Examples

sim.Q=c(500, 1001, 1500, 1998, 2507, 2994, 3498, 4006, 4511, 4991)

SalMod(Q.total=sim.Q)

#>  [1] 7.1506330 2.9165776 1.1938709 0.4895743 0.0000000 0.0000000 0.0000000
#>  [8] 0.0000000 0.0000000 0.0000000